server/requirements.txt
asgiref==3.2.3  
Django==3.0    
django-cors-headers==3.2.1
django-filter==2.0.0
djangorestframework==3.10.3 
djangorestframework-simplejwt==4.4.0 
Markdown==3.1.1
mysqlclient==1.4.6  
Pillow==6.2.1  
pip==19.3.1 
pytz==2019.3 
setuptools==42.0.2 
sqlparse==0.3.0  
wheel==0.33.6 
whitenoise==5.0.1 